UPDATE - persistance fixes
As mentioned in the previous commit, I have now fixed the persistence issue so that when the app is re-run, votes in blockchain and user status remain the same.

Added some new functions in blockchain.py and app.py. Still there exist debug statements which can be removed at a later point.

Next time:
Users can see who voted for who. This needs to be fixed. The next goal is to improve blockchain explorer.
Also have a look at different complex methods that can be implemented into the program.