MUST READ
An addition to previous commit.

Check if the logic is working properly if the user can actually validate their vote.

Also there was an error where you would inspect the bc explorer and the screen would go inactive so u would have to go back. Although i think this is fixed but check.

Main point: Check if the actual validation is working.